requirements.txt
alabaster==0.7.7
Babel==2.2.0
cssselect==0.9.1
d==0.2.2
Django==1.9
django-appconf==1.0.1
django-extra-views==0.7.1
django-imagekit==3.3
django-jsonify==0.3.0
django-registration==2.0.3
django-widget-tweaks==1.4.1
docutils==0.12
Jinja2==2.8
lxml==3.5.0
Markdown==2.6.5
MarkupSafe==0.23
mysqlclient==1.3.7
pilkit==1.1.13
Pillow==3.0.0
Pygments==2.1
pyquery==1.2.9
pytz==2015.7
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.4
sphinx-rtd-theme==0.1.9
wheel==0.24.0
